Force an element to be collected and stored

selenider_element/selenider_elements objects are generally lazy, meaning they only collect the actual element in the DOM when absolutely necessary, and forget it immediately after. This is to avoid situations where the DOM changes after an element has been collected, resulting in errors and unreliable behaviour.

These functions force an element to be collected and stored, making it eager rather than lazy. This is useful when you are operating on the same element multiple times, since only collecting the element once will improve performance. However, you must be sure that the element will not change on the page while you are using it.

Usage

cache_element(x, timeout = NULL)

Arguments

x

For cache_element(), a selenider_element object. For cache_elements(), a selenider_elements object.

timeout

How long to wait for the element(s) to exist while collecting them.

Value

The same type as x. The result of cache_element()/cache_elements() can be used as a normal selenider_element/selenider_elements object.

Details

These functions do not make selenider elements permanently eager. Further sub-elements will not be cached unless specified.

For example, consider the following code:

s(".class1") |>
  html_parent() |>
  cache_element() |>
  html_element(".class2")

In this example, the parent of the element with class ".class1" will be cached, but the child element with class ".class2" will not.

See also

Examples

html <- "
<div>
<p id='specifictext'></p>
<button></button>
</div>
"

session <- minimal_selenider_session(html)
#> [1] "/tmp/Rtmpx6JCg9/file186e7fa5d968.html"

# Selecting this button may be slow, since we are using relative XPath selectors.
button <- s("#specifictext") |>
  html_siblings() |>
  html_find(has_name("button"))

# But we need to click the button 10 times!
# Normally, this would involve fetching the button from the DOM 10 times
click_button_10_times <- function(x) {
  lapply(1:10, \(unnused) click(x))
  invisible(NULL)
}

# But with cache_element(), the button will only be fetched once
cached_button <- cache_element(button)
#> Error in self$send_command(msg, callback = callback_, error = error_,     timeout = timeout_): Session 68E4D302F4662661494C0280948A6E2A is closed.

click_button_10_times(cached_button)
#> Error in eval(expr, envir, enclos): object 'cached_button' not found

# But the cached button is less reliable if the DOM is changing
execute_js_fn("x => { x.outerHTML = '<button></button>'; }", button)
#> Error in self$send_command(msg, callback = callback_, error = error_,     timeout = timeout_): Session 68E4D302F4662661494C0280948A6E2A is closed.

try(click(cached_button, timeout = 0.1))
#> Error in eval(expr, envir, enclos) : object 'cached_button' not found

# But the non-cached version works
click(button)
#> Error in self$send_command(msg, callback = callback_, error = error_,     timeout = timeout_): Session 68E4D302F4662661494C0280948A6E2A is closed.

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }