Driver options

chromote_options() and selenium_options() return a list of options that can be passed to the options argument of selenider_session().

chromote_options() allows you to control the creation of a chromote driver created using chromote::ChromoteSession$new().

selenium_options() allows you to control the creation of a selenium driver.

selenium_server_options() should be passed to the server_options argument of selenium_options(), allowing you to control the creation of the server using selenium::selenium_server().

selenium_client_options() should be passed to the client_options argument of selenium_options(), allowing you to control the creation of a Selenium client created using selenium::SeleniumSession$new().

Usage

chromote_options(
  headless = TRUE,
  parent = NULL,
  width = 992,
  height = 1323,
  targetId = NULL,
  wait_ = TRUE,
  auto_events = NULL
)

selenium_options(
  client_options = selenium_client_options(),
  server_options = selenium_server_options()
)

selenium_server_options(
  version = "latest",
  host = "localhost",
  port = 4444L,
  selenium_manager = TRUE,
  interactive = FALSE,
  stdout = NULL,
  stderr = NULL,
  verbose = FALSE,
  temp = TRUE,
  path = NULL,
  extra_args = c(),
  ...
)

selenium_client_options(
  port = 4444L,
  host = "localhost",
  verbose = FALSE,
  capabilities = NULL,
  request_body = NULL,
  timeout = 60
)

Arguments

headless

Whether to run the browser in headless mode, meaning that you won't actually be able to see the browser as you control it. For debugging purposes and interactive use, it is often useful to set this to FALSE.

parent

The parent chromote session.

width, height, targetId, wait_, auto_events

Passed into chromote::ChromoteSession$new().

client_options

A selenium_client_options() object.

server_options

A selenium_server_options() object, or NULL if you don't want one to be created.

version

The version of Selenium server to use.

host

The host to use.

port

The port number to use.

selenium_manager, interactive, stdout, stderr, verbose, temp, path, extra_args, ...

Passed into selenium::selenium_server().

capabilities, request_body, timeout

Passed into selenium::SeleniumSession$new().