Deprecated functions

These functions are deprecated and will be removed in a future release. Use the options argument to selenider_session() instead. If you want to manually create a chromote or selenium session, use chromote::ChromoteSession, selenium::SeleniumSession and selenium::selenium_server() manually, since these functions are only a thin wrapper around them.

Usage

create_chromote_session(parent = NULL, ...)

create_selenium_server(
  browser,
  version = "latest",
  driver_version = "latest",
  port = 4444L,
  quiet = TRUE,
  selenium_manager = TRUE,
  ...
)

create_selenium_client(browser, port = 4444L, host = "localhost", ...)

create_rselenium_client(browser, port = 4444L, ...)

Arguments

parent, ..., version, driver_version, port, quiet, host

See the documentation for chromote_options(), selenium_options(), selenium_client_options(), wdman_server_options(), selenium_client_options() and rselenium_client_options() for details about what these arguments mean.

browser

The browser to use.

selenium_manager

If this is FALSE, wdman::selenium() will be used instead of selenium::selenium_server(). The equivalent of using wdman_server_options() over selenium_server_options() in selenium_options().

Value

create_chromote_session() returns a chromote::ChromoteSession object.

create_selenium_server() returns a processx::process or wdman equivalent.

create_selenium_client() returns a selenium::SeleniumSession object.

create_rselenium_client() returns an RSelenium::remoteDriver object.