Get the element associated with a selenider element

Turn a lazy selenium element or element collection into a backendNodeId (chromote) or an RSelenium::webElement. Use this to perform certain actions on the element that are not implemented in selenider.

get_actual_element() turns a selenider_element object into a single backendNodeId or RSelenium::webElement object. The function will wait for the object to exist in the DOM.

get_actual_elements() turns a selenider_elements object into a list of RSelenium::webElement objects, waiting for any parent objects to exist in the DOM.

Usage

get_actual_element(x, timeout = NULL)

get_actual_elements(x, timeout = NULL)

Arguments

x

A selenider_element or selenider_elements object, produced by find_element() / find_elements().

timeout

The timeout to use while asserting that the item exists. If NULL, the timeout of the selenider_element will be used.

Value

An integer (backendNodeId), or an RSelenium::webElement object. get_actual_elements() returns a list of such objects.

See also

Examples

html <- "
<div>
<p>Text</p>
<p>More text</p>
</div>
"

session <- minimal_selenider_session(html)

elem <- s("div") |>
  get_actual_element()

# The webDriver/ChromoteSession can be accessed using session$driver
driver <- session$driver

if (inherits(driver, "ChromoteSession")) {
  driver$DOM$getBoxModel(backendNodeId = elem)
} else {
  elem$getElementLocation()
}

elems <- ss("p") |>
  get_actual_elements()

if (inherits(driver, "ChromoteSession")) {
  driver$DOM$describeNode(backendNodeId = elems[[1]])
} else {
  elems[[1]]$describeElement()
}

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }