Hover over an element

Move the mouse over to an HTML element and hover over it, without actually clicking or interacting with it.

Usage

hover(x, js = FALSE, timeout = NULL)

Arguments

x

A selenider_element object.

js

Whether to hover the element using JavaScript.

timeout

How long to wait for the element to exist.

Value

x, invisibly

See also

Examples

html <- "
<button onmouseover = settext()> </button>
<p class = 'text'></p>
"

js <- "
function settext() {
  document.getElementsByClassName('text').item(0).innerHTML = 'Button hovered!'
}
"

session <- minimal_selenider_session(html, js = js)
#> [1] "/tmp/RtmpNpRp9O/file183de5b97bf.html"

html_expect(s(".text"), has_exact_text(""))
#> Error in html_expect(s(".text"), has_exact_text("")): Condition failed after waiting for 4 seconds:
#> `has_exact_text("")`
#> ℹ `x` does not have exact text "".
#> ℹ Actual text: .
#> Caused by error in `has_exact_text()`:
#> ! `x` does not exist in the DOM.

s("button") |>
  hover()
#> Error in hover(s("button")): To hover over `x`, it must exist.
#> ℹ After 4 seconds, `x` was not present.

html_expect(s(".text"), has_text("Button hovered!"))
#> Error in html_expect(s(".text"), has_text("Button hovered!")): Condition failed after waiting for 4 seconds:
#> `has_text("Button hovered!")`
#> ℹ `x` does not have text "Button hovered!".
#> ℹ Actual text: .
#> Caused by error in `has_text()`:
#> ! `x` does not exist in the DOM.

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }