Extract a subset of HTML elements

Operators to extract a subset of elements, or a single element, from a selenider element collection.

html_filter() and html_find() allow you to use conditions to filter HTML elements (see is_present() and other conditions). html_find() returns the first element that satisfies one or more conditions, while html_filter() returns every element that satisfies these conditions.

[ and [[ with a numeric subscript can be used on an element collection to filter the elements by position. [ returns a single element at a specified location, while [[ returns a collection of the elements at more than one position.

Usage

html_filter(x, ...)

html_find(x, ...)

# S3 method for selenider_elements
[(x, i)

# S3 method for selenider_elements
[[(x, i)

Arguments

x

A selenider_elements object.

...

<dynamic-dots> Conditions (functions or function calls) that are used to filter the elements of x.

i

A number (or for [, a vector of one or more numbers) used to select elements by position.

Value

html_filter() and [ return a selenider_elements object, since they can result in multiple elements. html_find() and [[ return a single selenider_element object.

Details

As with the html_element() and html_elements() functions, these functions are lazy, meaning that the elements are not fetched and filtered until they are needed.

Conditions can be functions or function calls (see html_expect() for more details).

See also

Examples

html <- "
<button disabled>Button 1</button>
<button>Button 2</button>
<p>Text</p>
<div style='display:none;'></div>
"
session <- minimal_selenider_session(html)
#> [1] "/tmp/RtmpGbjLc8/file186bf8a560d.html"

elements <- ss("./*")

# Gives the same result as s()
elements[[1]]
#> A selenider element selecting:
#> The first element with css selector "./*".

elements[1:3]
#> A collection of selenider elements selecting:
#> The 1st, 2nd, and 3rd elements with css selector "./*".

elements[-2]
#> A collection of selenider elements selecting:
#> All elements with css selector "./*" except the second.

elements |>
  html_filter(is_visible)
#> A collection of selenider elements selecting:
#> • The elements with css selector "./*" matching the following condition:
#>   `is_visible`

elements |>
  html_find(is_visible)
#> A selenider element selecting:
#> • The first element with css selector "./*" matching the following condition:
#>   `is_visible`

# The above is equivalent to:
visible_elems <- elements |>
  html_filter(is_visible)
visible_elems[[1]]
#> A selenider element selecting:
#> • The first element with css selector "./*" matching the following condition:
#>   `is_visible`

# In R >= 4.3.0, we can instead do:
# ss(".class1") |>
#   html_filter(is_visible) |>
#   _[[1]]

ss("button") |>
  html_filter(is_enabled)
#> A collection of selenider elements selecting:
#> • The elements with css selector "button" matching the following condition:
#>   `is_enabled`

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }