Iterate over an element collection

Use html_flatmap() when you want to select further sub-elements for each element of a collection.

html_flatmap() allows you to apply a function to each element of a selenider_elements object, provided that the function returns a selenider_element/selenider_elements object itself. The result will then be flattened into a single selenider_elements object. The benefit of this over traditional iteration techniques is that the laziness of the elements will be maintained, and nothing will be fetched from the DOM.

element_list() transforms a selenider_elements object into a list of selenider_element objects. The result can then be used in for loops and higher order functions like lapply()/purrr::map() (whereas a selenider_element object cannot)

html_flatmap() works by executing .f on a mock element, then recording the results in x. This means that no matter the length of x, .f is only evaluated once, and during the html_flatmap() call. For this reason, .f should not invoke any side effects or do anything other than selecting sub-elements.

element_list() essentially turns x into: list(x[[1]], x[[2]], ...) However, to do this, the length of x must be computed. This means that while each element inside the list is still lazy, the list itself cannot be considered lazy, since the number of elements in the DOM may change. To avoid problems, it is recommended to use an element list just after it is created, to make sure the list is an accurate representation of the DOM when it is being used.

Usage

html_flatmap(x, .f, ...)

element_list(x, timeout = NULL)

Arguments

x

A selenider_elements object.

.f

A function to apply to each element of x.

...

Passed into .f.

timeout

How long to wait for x to exist while computing its length.

Value

html_flatmap() returns a selenider_element object. element_list() returns a list of selenider_element objects.

Examples

html <- "
<div id='div1'>
  <p>Text 1</p>
</div>
<div id='div2'>
  <p>Text 2</p>
</div>
<div id='div3'>
  <p>Text 3</p>
</div>
<div id='div4'>
  <p>Text 4</p>
</div>
"

session <- minimal_selenider_session(html)
#> [1] "/tmp/RtmpLZl2AD/file18833ff9731b.html"

divs <- ss("div")

# Get the <p> tag inside each div.
divs |>
  html_flatmap(\(x) x |> html_element("p"))
#> A collection of selenider elements selecting:
#> • The elements with css selector "div"
#> • The first child of each element with css selector "p"

# Or:
p_tags <- divs |>
  html_flatmap(html_element, "p")

# To get the text in each tag, we can't use html_flatmap()
for (elem in element_list(p_tags)) {
  print(html_text(elem))
}
#> list()

# Or:
lapply(element_list(p_tags), html_text)
#> list()
#> list()

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }