Combine multiple HTML elements

Combine a set of selenider_element/selenider_elements objects into a single selenider_elements object, allowing you to perform actions on them at once. c() and html_flatten() do the same thing, but html_flatten() works on lists of selenider_element/selenider_elements objects.

Usage

html_flatten(...)

# S3 method for selenider_element
c(...)

# S3 method for selenider_elements
c(...)

Arguments

...

<dynamic-dots> selenider_element or selenider_elements objects to be combined, or lists of such objects.

Value

A selenider_elements object.

Examples

html <- "
<div id='id1'></div>
<div class='.class2'></div>
<button id='button1'>Click me!</button>
<div class='button-container'>
  <button id='button2'>No, click me!</button>
</div>
"

session <- minimal_selenider_session(html)
#> [1] "/tmp/RtmpPR9C0R/file17f9e259355.html"

button_1 <- s("#button1")
button_2 <- s("#button2")

buttons <- html_flatten(button_1, button_2)

buttons |>
  html_expect_all(is_enabled)

buttons |>
  element_list() |>
  lapply(click)
#> list()

# Doesn't just have to be single elements
first_2_divs <- ss("div")[1:2]

html_flatten(first_2_divs, button_2) |>
  length()
#> [1] 0

# We would like to use multiple css selectors and combine the results
selectors <- c(
  "#id1", # Will select 1 element
  "button", # Will select 2 elements
  "p" # Will select 0 elements
)

lapply(selectors, ss) |>
  html_flatten() |>
  length() # 3
#> [1] 0

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }