Open a URL

Navigate the browser to specified URL, waiting until the page is considered open before finishing.

Usage

open_url(url, session = NULL)

Arguments

url

The URL to navigate to: a string.

session

A `selenider_session object. If not specified, the global session object (the result of get_session()) is used.

Value

The session object, invisibly.

See also

Examples

session <- selenider_session()

open_url("https://r-project.org")

# Or:
open_url(session = session, "https://r-project.org")
#> Error in private$register_event_listener("Page.loadEventFired", callback_,     timeout_): Session FEEDC49AD807522C58D84EB3CBBD5CFC is closed.

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }