Quickly select HTML elements

Both s() and ss() allow you to select elements without specifying a session object.

s() selects a single element, being a shorthand for find_element() on the current session.

ss() selects multiple elements, being a shorthand for find_elements().

Usage

s(
  css = NULL,
  xpath = NULL,
  id = NULL,
  class_name = NULL,
  name = NULL,
  link_text = NULL
)

ss(
  css = NULL,
  xpath = NULL,
  id = NULL,
  class_name = NULL,
  name = NULL,
  link_text = NULL
)

Arguments

css

A css selector.

xpath

An XPath.

id

The id of the element you want to select.

class_name

The class name of the element you want to select.

name

The name attribute of the element you want to select.

link_text

The link text of the link element that you would like to select.

Value

s() returns a selenider_element object. ss() returns a selenider_elements object.

Details

Both functions allow the starting point for chains of selectors to be made more concise. Both use get_session() to get the global session object.

See also

Examples

html <- "
<div>
<p id='id1' class='inner'></p>
<div class='child'>
<p class='inner'></p>
</div>
</div>
"

session <- minimal_selenider_session(html)
#> [1] "/tmp/RtmpZ9Ij1H/file188e35756701.html"

s("#id1")
#> A selenider element selecting:
#> The first element with css selector "#id1".

# This is the equivalent of:
find_element(session, "id1")
#> A selenider element selecting:
#> The first element with css selector "id1".

ss(".inner")
#> A collection of selenider elements selecting:
#> The elements with css selector ".inner".

# This is the equivalent of:
find_element(session, ".inner")
#> A selenider element selecting:
#> The first element with css selector ".inner".

# This provides a more concise way to begin a chain of selectors
s("div") |>
  find_element(".child") |>
  find_element(".inner")
#> A selenider element selecting:
#> • The first element with css selector "div".
#> • The first child element with css selector ".child".
#> • The first child element with css selector ".inner".

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }