Compatibility with rvest

These functions ensure compatibility with rvest and xml2.

selenider_prefer() ensures that all functions common to selenider and rvest (e.g. html_attr()) work on both selenider elements and rvest node objects. conflicted is used to resolve function conflicts.

selenider allows xml2::read_html() and rvest::html_table() to work on selenider_element and selenider_session objects.

Usage

selenider_prefer(quiet = TRUE)

read_html.selenider_session(
  x,
  encoding = "",
  ...,
  options = c("RECOVER", "NOERROR", "NOBLANKS")
)

read_html.selenider_element(
  x,
  encoding = "",
  timeout = NULL,
  outer = TRUE,
  ...,
  options = c("RECOVER", "NOERROR", "NOBLANKS")
)

html_table.selenider_session(
  x,
  header = NA,
  trim = TRUE,
  fill = deprecated(),
  dec = ".",
  na.strings = "NA",
  convert = TRUE
)

html_table.selenider_element(
  x,
  header = NA,
  trim = TRUE,
  fill = deprecated(),
  dec = ".",
  na.strings = "NA",
  convert = TRUE
)

Arguments

quiet

If TRUE, output from conflicted::conflict_prefer() will be suppressed.

x

A selenider_session/selenider_element object.

encoding, ..., options

Passed into xml2::read_html().

timeout

How long to wait for x to exist in the DOM before throwing an error.

outer

Whether to read the inner (all children of the current element) or outer (including the element itself) HTML of x.

header, trim, fill, dec, na.strings, convert

Passed into rvest::html_table().

Value

read_html() returns an XML document. html_table() returns a tibble or list of tibbles (when applied to a page).

Examples

library(rvest)
#> 
#> Attaching package: ‘rvest’
#> The following objects are masked from ‘package:selenider’:
#> 
#>     back, forward, html_attr, html_attrs, html_children, html_element,
#>     html_elements, html_name, html_text
selenider_prefer(quiet = FALSE)
#> [conflicted] Will prefer selenider::html_attr over any other package.
#> [conflicted] Will prefer selenider::html_attrs over any other package.
#> [conflicted] Will prefer selenider::html_element over any other package.
#> [conflicted] Will prefer selenider::html_elements over any other package.
#> [conflicted] Will prefer selenider::html_name over any other package.
#> [conflicted] Will prefer selenider::html_text over any other package.

html <- "
<div>
<p>Example text</p>
</div>

<table>
  <tr>
    <th>Column 1</th>
    <th>Column 2</th>
  </tr>
  <tr>
    <td>10</td>
    <td>Data</td>
  </tr>
</table>
"

session <- minimal_selenider_session(html)
#> [1] "/tmp/RtmphzEqOh/file182c8504749.html"

read_html(session)
#> Error in self$send_command(msg, callback = callback_, error = error_,     timeout = timeout_): Session C84F2D2B98A7D0991CD2325885565CC2 is closed.
read_html(s("div"))
#> {html_document}
#> <html>
#> [1] <body><div class="container page">\n      <div class="row">\n        <div ...

html_table(s("table"))
#> Error in get_element_for_action(x, action = action, conditions = list(),     timeout = timeout, failure_messages = c(), conditions_text = c(),     call = call): To Read the HTML of `x`, it must exist.
#> ℹ After seconds, `x` was not present.

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }