Start a session

Begin a session in selenider, setting the session globally unless otherwise specified.

Usage

selenider_session(
  browser = c("chrome", "firefox", "phantomjs", "internet explorer"),
  timeout = 4,
  driver = NULL,
  local = TRUE,
  quiet = TRUE,
  .env = rlang::caller_env()
)

Arguments

browser

The name of the browser to run the session in.

timeout

The default time to wait when collecting an element.

driver

A driver object to use instead of creating one manually.

local

Whether to set the session as the local session object, using local_session().

quiet

Whether to let RSelenium::rsDriver() display messages. By default, this output is suppressed, as it is not usually useful.

.env

Passed into local_session() function, to define the environment in which the session is used. Change this if you want to create the session inside a function and then use it outside the function.

Value

A selenider_session object.

Details

selenider_session() uses RSelenium::rsDriver() to create a browser session. It then uses withr to set the session locally (unless otherwise specified). To reset the local session object, use withr::deferred_run().

See also

Examples

if (FALSE) {
# Note that these examples will create real browser sessions, so should only be
# run interactively.

session <- selenider_session(browser = "firefox")

withr::deferred_run() # Reset the session

# If we want to use the session manually:
selenider_session(local = FALSE)

session_1 <- selenider_session(timeout = 10)
# session_1 is the local session here

my_function <- function() {
  session <- selenider_session(browser = "internet explorer")

  # In here, session_2 is the local session
  get_session()
} # When the function finishes executing, the session is closed

my_function() # Returns `session_2`

# But outside the function, session_1 is the global session again

get_session() # Returns `session_1`
 
withr::deferred_run() # Close `session_1`

# If we want to use a session outside the scope of a function,
# we need to use the `.env` argument.
create_session <- function(timeout = 10, .env = rlang::caller_env()) {
  # caller_env() is the environment where the function is called
  selenider_session(browser = "firefox", timeout = timeout, .env = .env)
}

my_session <- create_session()

# We can now use this session outside the `create_session()` function
get_session()

# `my_session` will be closed automatically.
}