Driver options

chromote_options() and selenium_options() return a list of options that can be passed to the options argument of selenider_session().

chromote_options() allows you to control the creation of a chromote driver created using chromote::ChromoteSession$new().

selenium_options() allows you to control the creation of a selenium driver.

selenium_server_options() and wdman_server_options() should be passed to the server_options argument of selenium_options(). By default, the former is used, meaning that the server is created using selenium::selenium_server(). If wdman_server_options() is used instead, the server will be created using wdman::selenium().

selenium_client_options() should be passed to the client_options argument of selenium_options(), allowing you to control the creation of a Selenium client created using selenium::SeleniumSession$new().

Instead of using selenium_client_options(), you can use rselenium_client_options() to control the creation of an RSelenium::remoteDriver() object instead. This is not recommended, since RSelenium is incompatible with newer versions of Selenium.

Usage

chromote_options(
  headless = TRUE,
  parent = NULL,
  width = 992,
  height = 1323,
  targetId = NULL,
  wait_ = TRUE,
  auto_events = NULL
)

selenium_options(
  client_options = selenium_client_options(),
  server_options = selenium_server_options()
)

selenium_server_options(
  version = "latest",
  port = 4444L,
  selenium_manager = NULL,
  verbose = FALSE,
  temp = TRUE,
  path = NULL,
  interactive = FALSE,
  echo_cmd = FALSE,
  extra_args = c()
)

wdman_server_options(
  version = "latest",
  driver_version = "latest",
  port = 4444L,
  check = TRUE,
  verbose = FALSE,
  retcommand = FALSE,
  ...
)

selenium_client_options(
  port = 4444L,
  host = "localhost",
  verbose = FALSE,
  capabilities = NULL,
  request_body = NULL,
  timeout = 20
)

rselenium_client_options(
  port = 4444L,
  host = "localhost",
  path = "/wd/hub",
  version = "",
  platform = "ANY",
  javascript = TRUE,
  native_events = TRUE,
  extra_capabilities = list()
)

Arguments

headless

Whether to run the browser in headless mode, meaning that you won't actually be able to see the browser as you control it. For debugging purposes and interactive use, it is often useful to set this to FALSE.

parent

The parent chromote session.

width, height, targetId, wait_, auto_events

Passed into chromote::ChromoteSession$new().

client_options

A selenium_client_options() object.

server_options

A selenium_server_options() or wdman_server_options() object.

version

The version of Selenium server to use.

port

The port number to use.

selenium_manager, verbose, temp, path, interactive, echo_cmd, extra_args

Passed into selenium::selenium_server().

driver_version

The version of the browser-specific driver to use.

check, retcommand, ...

Passed into wdman::selenium().

host, capabilities, request_body, timeout

Passed into selenium::SeleniumSession$new().

platform, javascript, native_events, extra_capabilities

Passed into RSelenium::remoteDriver().