Open a URL

Navigate the browser to specified URL, waiting until the page is considered open before finishing.

Usage

open_url(url, timeout = NULL, session = NULL)

Arguments

url

The URL to navigate to: a string.

timeout

The maximum time to wait for the page to load, in seconds. This defaults to 60, unless in a Github Action, in which case it defaults to 5 minutes.

session

A selenider_session object. If not specified, the global session object (the result of get_session()) is used.

Value

The session object, invisibly.

See also

Examples

session <- selenider_session()

open_url("https://r-project.org")

# Or:
open_url(session = session, "https://r-project.org")