Start a session

Begin a session in selenider, setting the session globally unless otherwise specified.

create_chromote_session(), create_selenium_client() and create_selenium_server() are low-level functions that allow more control over making a web driver, which can then be passed into the driver argument to selenider_session().

Usage

selenider_session(
  session = NULL,
  browser = NULL,
  view = FALSE,
  timeout = 4,
  driver = NULL,
  local = TRUE,
  quiet = TRUE,
  .env = rlang::caller_env()
)

create_chromote_session(...)

create_selenium_server(
  browser,
  version = "latest",
  port = 4567L,
  quiet = TRUE,
  ...
)

create_selenium_client(browser, port = 4567L, ...)

Arguments

session

The package to use as a backend: either "chromote" or "selenium". By default, chromote is used, since this tends to be faster and more reliable.

browser

The name of the browser to run the session in; one of "chrome", "firefox", "phantomjs" or "internet explorer" (only on Windows). IF NULL, the function will try to work out which browser you have installed. If we are using chromote, this option is ignored, since chromote only works on Chrome.

view

Whether to open the browser and view it, for visual testing. This is ignored if session is "selenium", since selenium drivers cannot be headless.

timeout

The default time to wait when collecting an element.

driver

A driver object to use instead of creating one manually. This can be one of:

A chromote::ChromoteSession object (the result of create_chromote_session()).

A shinytest2::AppDriver object.

An RSelenium::remoteDriver() object (the result of create_selenium_client()).

A Selenium server object (the result of wdman::selenium(), or create_selenium_server()). In this case, the client object will be created using the server object.

A list/environment containing the RSelenium::remoteDriver() object, the Selenium server object, or both. See Details for more information about providing a custom driver object.

local

Whether to set the session as the local session object, using local_session().

quiet

Whether to let RSelenium::rsDriver() display messages. By default, this output is suppressed, as it is not usually useful. Chromote does not display any output when creating a session.

.env

Passed into local_session() function, to define the environment in which the session is used. Change this if you want to create the session inside a function and then use it outside the function.

...

Arguments to finetune the creation of the specific driver.

For create_chromote_session(), these are passed into chromote::ChromoteSession$new().

For create_selenium_server(), these are passed into wdman::selenium().

For create_selenium_server(), these are passed into RSelenium::remoteDriver().

version

The version of the webdriver (chromedriver, geckodriver, etc.) to use. You should only need to change this if Chrome is being used, as the version of the driver depends on the version of Chrome.

port

The port to run Selenium on.

Value

A selenider_session object. Use session$driver to retrieve the driver object that controls the browser.

Structure of a selenider session

A selenider_session object is an S3 list, meaning its properties can be accessed using $. Most notably, using session$driver allows access to the driver object which actually controls the browser. If you are using Selenium, use session$driver$client to access the remoteDriver object. These objects are useful if you want to do something with the driver that is not directly supported by selenider. See get_actual_element() for some examples of this.

Custom drivers

Custom driver objects are good if you want more low-level control over the underlying functions that create the webdrivers that actually control the browser. However, it is recommended to use the selenider functions (e.g. create_selenium_client()) over wdman::selenium() for better error messages and more reliable behaviour.

TODO: Add vignette references when available

Chromote

Supplying a custom chromote::ChromoteSession object can allow you to manage the underlying chromote::Chromote process that is used to spawn sessions. For example:

my_chromote_object <- chromote::Chromote$new()

session <- selenider_session(
  driver = create_chromote_session(parent = my_chromote_object)
)

You can also supply a shinytest2::AppDriver object, allowing selenider and shinytest2 to share a session:

shiny_app <- shiny::shinyApp(
  ui = shiny::fluidPage(
    # ... Your UI
  ),
  server = function(input, output) {
    # ... Your server
  }
)

app <- shinytest2::AppDriver$new()

session <- selenider_session(
  driver = app
)

RSelenium

If you want to manually create both the client and the server, you can do the equivalent of the following:

session <- selenider_session(
  driver = list(
    client = create_selenium_client("chrome"),
    server = create_selenium_server("chrome")
  )
)

However, it can sometimes be useful to omit the server, for example when you are running the Selenium server using Docker. In this case, you need to make sure the ip and port are matched correctly.

session <- selenider_session(
  driver = create_selenium_client(
    remoteServerAddr = "<IP ADDRESS>",
    port = 1234L
  )
)

See also

Examples

session_1 <- selenider_session(timeout = 10)
# session_1 is the local session here

get_session() # Returns session 1
#> A selenider session

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }

my_function <- function() {
  session_2 <- selenider_session()

  # In here, session_2 is the local session
  get_session()
  # \dontshow{
  # Clean up all connections and invalidate default chromote object
  selenider_cleanup()
  # }
} # When the function finishes executing, the session is closed

my_function() # Returns `session_2`

# If we want to use a session outside the scope of a function,
# we need to use the `.env` argument.
create_session <- function(timeout = 10, .env = rlang::caller_env()) {
  # caller_env() is the environment where the function is called
  selenider_session(timeout = timeout, .env = .env)
}

my_session <- create_session()

# We can now use this session outside the `create_session()` function
get_session()
#> A selenider session

# `my_session` will be closed automatically.

# \dontshow{
# Clean up all connections and invalidate default chromote object
selenider_cleanup()
# }